backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
dj-database-url==2.3.0
Django==5.2
django-channels==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-decouple==3.8
requests==2.32.4
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.13.2
urllib3==2.5.0
whitenoise==6.9.0
zope.interface==7.2
